refactor(#792): decompose compact_chain_block_handlers::wire (584-line fn) by responsibility
Orchestrator + 4 helpers: wire_block_toggle_and_model, wire_chain_toggle (base) +
wire_block_delete, wire_block_reorder (compact_chain_block_delete.rs sibling). Same
closures/captures via &Ctx. File 628->462, off debt. compact tests: 25 pass, zero warnings.
