+++ crates/adapter-gui/src/lib.rs
@@ -47,6 +47,7 @@ mod compact_block_layout;
 mod compact_block_tabs;
 mod compact_block_view;
 mod compact_chain_block_handlers;
+mod compact_chain_block_delete;
 pub mod compact_chain_callbacks;
 mod compact_chain_delete_wiring;
 mod compact_chain_header_wiring;
